pyproject.toml
[tool.poetry]
name = "farabin-api"
version = "0.1.0"
description = "Farain Backend"
authors = ["ahMAD ASSadi <ds1380ahmad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
amqp = "5.3.1"
asgiref = "3.8.1"
attrs = "24.2.0"
billiard = "4.2.1"
celery = {version = "5.4.0", extras = ["redis"]}
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
contourpy = "1.3.1"
cron-descriptor = "1.4.5"
cycler = "0.12.1"
django = "^5.1.9"
django-admin-extensions = "0.10.0"
django-admin-interface = "0.29.2"
django-celery-beat = "2.7.0"
django-colorfield = "0.11.0"
django-cors-headers = "4.6.0"
django-environ = "0.11.2"
django-nested-admin = "4.1.1"
django-timezone-field = "7.0"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "0.27.2"
flower = "2.0.1"
fonttools = "4.55.0"
humanize = "4.11.0"
idna = "3.10"
inflection = "0.5.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
kiwisolver = "1.4.7"
kombu = "5.4.2"
msgpack = "1.1.0"
numpy = "2.1.3"
packaging = "24.2"
pdf2image = "1.17.0"
pillow = "11.0.0"
prometheus-client = "0.21.0"
prompt-toolkit = "3.0.48"
psycopg2-binary = "2.9.10"
pyjwt = "2.10.1"
pyparsing = "3.2.0"
python-crontab = "3.2.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-magic = "0.4.27"
python-monkey-business = "1.1.0"
python-slugify = "8.0.4"
pytz = "2024.2"
pyyaml = "6.0.2"
redis = "5.2.0"
referencing = "0.35.1"
requests = "2.32.3"
rest-framework-simplejwt = "0.0.2"
rpds-py = "0.21.0"
six = "1.16.0"
sqlparse = "0.5.2"
text-unidecode = "1.3"
toposort = "1.10"
tornado = "6.4.2"
typing-extensions = "4.12.2"
tzdata = "2024.2"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
vine = "5.1.0"
waitress = "3.0.2"
wcwidth = "0.2.13"
whitenoise = "6.8.2"
pydantic = "^2.10.4"
cohere = "^5.13.4"
openai = "^1.58.1"
environ = "^1.0"
django-redis = "^5.4.0"
supervisor = "^4.2.5"
networkx = "^3.4.2"
graphviz = "^0.20.3"
pygraphviz = "^1.14"
sqlalchemy = "^2.0.37"
openpyxl = "^3.1.5"
drf-nested-routers = "^0.94.1"
gunicorn = "^23.0.0"
pandas = "^2.2.3"
django-debug-toolbar = "^5.0.1"
locust = "^2.33.1"
google-auth = "^2.39.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.167.0"
pytest = "^8.3.5"
pytest-django = "^4.11.1"
django-lifecycle = "^1.2.4"
google-genai = "^1.14.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
